Can verify cleartext signatures, if ext-gnupg is loaded
The "green" flag as explained at https://www.gnupg.org/documentation/manuals/gpgme/Verify.html seems to be enough for this case and currently we don't care about any other info. But the flag is also set when the signing key is unknown (not in the keyring) so if that's the case, we also check the red flag is unset.