Don't suggest parsing the exception messages with Markdown etc.
The server could redirect for example /security.txt to /something?url=<HTML> (or to ?url=[dont click](https://phishing.example) and then HTML would be processed by the browser which could result in XSS for example.

Remove backticks from the exception messages which could also sort of suggest that you can Markdown-format them (reverts most of 182001a745b0102a04526aed53d5aa01c2b23ec1) and instead use placeholders wherever the value could be formatted as <code>. The user of the library is then responsible to format the values properly and use them to replace the placeholders.

What I'm going to do in one of the sites is something similar to this:

```php
$html = new Html();
foreach (explode('%s', $message) as $key => $part) {
  $html->addText($part); // don't use addHtml here
  $code = new Html('code')->addText($values[$key]); // not even here
  $html->addHtml($code);
}
$html->addText($parts[$i]); // adding the final part
```